Implement own check procedure for required CLI arguments
The required option in argparse can not be used without beeing in
conflict with the usage of config files instead of CLI options.

This is now fixed and one needs to add required options in the
required_options_dict member variable for a specific command if it sould
be checked.